Using retries when committing files to the final destination
And if the file already exists, we fallback to not doing anything. Since
we're doing moves instead of copies, there's no way to move with
overwrite, the issue then becomes, what if the file that got moved is in
a failed/corrupted state. We could move to a copy-instead-of-move
approach, would have performnace impact plus could have other
repercussion I don't know about.